search k1Vfydu2 - update dispute_date txn search specs
Elasticsearch is now the source of truth for dispute search. The
existing integration specs for dispute_date relied on disputed
transactions that were created during dbdo, but this is no longer
possible since this data cannot be backfilled into ES. Instead, we must
create a new disputed transaction, which will go through the datapipe to
populate ES with the dispute information.

In revisiting the dispute_date specs, we found that some of them were
relying on a dispute_date search that was hardcoded to search for
disputes created at midnight UTC. This was masking a subtle bug in the
GW's SQL query for dispute_date, which has been fixed by removing the
SQL search path and exclusively querying Elasticsearch. This fix also
exposed a subtle difference between the logic of the SQL and
Elasticsearch dispute_date queries.

The logical difference responsible for fixing the bug also introduces a
small regression in an edge case when searching on dispute_date.
Namely, we would previously only return disputes when the date range
criteria spanned midnight UTC of the received_date. For example, a
dispute created on 3/8 would not be returned by a search from 3/8
00:00:01 to 3/8 23:59:59 because that range does not include midnight on
3/8. The new behavior is slightly more permissive, and would return the
dispute.

We feel that the benefit of fixing the bug outweighs any change to
undefined behavior exercised by this edge case, so we are simply
removing the assertions that no longer hold true.