Expose verification on success (#225)
### What
Expose the verification in the response when a verification succeeds.

### Why
Other SDKs expose the verification on the response when it succeeds, and
in this SDK when a transaction succeeds it is exposed on the response as
well. For consistency it makes sense to expose it, and the fact it is
not exposed restricts the types of tests we can write about the
verification returned when a verification succeeds.