Simplify test TCPSocket for SSL verification
This removes a "hack" method that forced tests to wait until a TCPSocket instance
was established. We can simply create the connection and test away. The hack was
originally added  for v1.0.1 more than ten years ago and I can't find a reason for
it anymore. Furthermore, it's now breaking when testing against ruby 2.7.