Pend GraphQL client test until Journalist can rotate their certs (#479)
Why This Change Is Necessary
========================================================================
This test is failing due to journalist having an expired cert (and the
test makes a query that hits journalist). Let's pend this so gateway PRs
are unstuck while journalist folks get things back in order.

What These Changes Do To Address the Issue
========================================================================
Pends one test. One small test that isn't even checking for content of a
reports, but the nature of the test is such that to generate a
validation error we need to talk to journalist. Sure, I could update the
test to do some other query that would always result in a validation
error, but this is the path of lesser resistance. Maybe we just end up
removing it all together?

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

   * [Slack Conversation][3]

------------------------------------------------------------------------

[3]: https://paypal.slack.com/archives/C5LSJH970/p1690840512891979