OPENDEV- update rspec2 to 3 syntax (#487)
* OPENDEV- use transpec to update rspec 2 to 3 syntax

Why This Change Is Necessary
========================================================================
For good code cleanliness, we should be running tests on the latest
syntactical version of rspec, 3.

What These Changes Do To Address the Issue
========================================================================
Used the transpec gem to update all outdated test syntax in the library.
Note, transpec hasn't been updated in a while and I initially
encountered issues with bundler and gem version conflicts and RVM
install/execution paths. To get around this, I hosed the original
Dockerfile and did a vanilla ruby + bundler install with no rvm, then
installed a special version of bundler that would run with
transpec(1.17.3). Then I removed the Gemfile.lock and removed from the
Gemfile the rubocop dependency so that transpec could install.

Kudos to James Coleman for being by sounding board on this work!

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

------------------------------------------------------------------------