OPENDEV SDK lint rules in parity with gateway rules, part 2  (#499)
* OPENDEV SDK lint rules in parity with gateway rules, part 2

Why This Change Is Necessary
========================================================================
When we first implemented rubocop in the SDK, we copy pasted the rules
employed by the gateway repo. It's been a few years since we last got
the SDK's rules in sync with that codebase, so this commit starts that
work with relevant linting cops in alphabetical order (f - p).

What These Changes Do To Address the Issue
========================================================================
Adds linting rules in alphabetical order from the gateway rubocop YAML
file to the SDK file, and stopping at the first rule that was violated
(Lint/ParenthesesAsGroupedExpression). Fixed that linting issue and
pushed up this work for review.

There were some rules in the gateway file that had been set to false and
so weren't being run. I opted not to include those in the SDK file and
stick rules linting rules that were _on_.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

------------------------------------------------------------------------

* PR feedback fixup a-z sorting