OPENDEV: Nodoc nomore (#492)
* OPENDEV Remove old Rdoc syntax

Why This Change Is Necessary
========================================================================
A long time ago, we used to publish [Rdocs][1] as part of our regular
release process. We don't do that anymore now that we have a separate
developer docs site with far better reference documentation.

Part of rdoc syntax is to add the comment " :nodoc:" if you didn't want
a given class or function documented. Why continue to have these
comments (and the bits of space they take) if we don't use them?

What These Changes Do To Address the Issue
========================================================================
Removes "nodoc" comments. There are a handful of places where I added
some addition NEXT_MAJOR_VERSION notes for us to explore a refactor, or
a nearby redundant comment was deleted.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

------------------------------------------------------------------------

[1]: https://ruby.github.io/rdoc/

* OPENDEV remove redundant comments

Why This Change Is Necessary
========================================================================
The source of truth is the developer docs and their reference pages. We
should remove superfluous comments so that devs have less of a crutch
and truly leverage the reference docs.

What These Changes Do To Address the Issue
========================================================================
Removes comments that are redundant or better elaborated in the dev
docs.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

------------------------------------------------------------------------

* OPENDEV add a few NEXT_MAJOR_VERSION comments for exploration

Why This Change Is Necessary
========================================================================
While I was working in the SDK I noticed some things that we can only
visit when doing major version changes because actions would be breaking
changes. Better to add a comment so we don't forget about it.

What These Changes Do To Address the Issue
========================================================================
Adds a few NEXT_MAJOR_VERSION comments with my thoughts for when we can
address them

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant Links:**

------------------------------------------------------------------------