Pend the one test blocking all of braintree
This spec that we're `xit`-ing tests that Atmosphere request are
valid when given inputs, and it happens to reach out to the Gateway for
that request. Due to a CA expiration and the fact that atmosphere grabs
the CA from secrets vault, there's currently a mismatch in CAs between
the one committed in the Gateway repo and Atmosphere.

While we resolve the dev CA mismatch, we're `xit`-ing this spec to
unblock the company.