Fix merchant creation specs (#548)
* Fix merchant creation specs

Co-authored-by: Tiago Motta Jorge <tmottajorge@paypal.com>

* Move to using GBR merchants in other specs, fix which currency is default

There were a few tests in other files where we were still attempting to
create USA merchants to test particular functionality. Also, the
remaining test failures in `merchant_spec` had to do with the fact that
USD was assumed to be the default currency, when in fact GBP would be
the new default for our merchants. The EU specific test already expected
this, and the US based test is removed because we no longer need to test
those expectations.

* Fix a failing transaction integration spec.

After getting the Gateway PR merged, the CI run revealed one last spot
where we were not creating a merchant with a proper country.

---------

Co-authored-by: Rajaram Gurumurthi <rajaram.gurumurthi@paypal.com>
Co-authored-by: Matt Hogan <mjhogan@paypal.com>